fixes controller disconnection. Adds some debug prints
Former-commit-id: ac240980a9d4039c4923b5a65a32b11f3315847c

+++ src/entity.cpp
@@ -122,7 +122,7 @@ static void LoadEntityData(FILE* File, entity* Entity, game_state* GameState, b3
         }
         else if(StartsWith(&LineBuffer[0], "health"))
         {
-            sscanf(LineBuffer, "health %hd", &Entity->FullHealth);
+            sscanf(LineBuffer, "health %d", &Entity->FullHealth);
             Entity->Health = Entity->FullHealth;
         }
         else if(StartsWith(&LineBuffer[0], "animation"))
@@ -219,6 +219,7 @@ static void LoadEntityData(FILE* File, entity* Entity, game_state* GameState, b3
                    &NewPointLight.Type ,&NewPointLight.Active,
                    &NewPointLight.Pointlight.Radius, &NewPointLight.Pointlight.Intensity,&NewPointLight.Color.x,&NewPointLight.Color.y,&NewPointLight.Color.z,&NewPointLight.Color.w, &NewPointLight.Pointlight.ConstantAtten, &NewPointLight.Pointlight.LinearAtten, &NewPointLight.Pointlight.ExponentialAtten);
             NewPointLight.Pointlight.Position = Entity->Position;
+            
             if(Entity->LightSourceHandle == -1)
             {
                 Entity->LightSourceHandle = GameState->LightSourceCount;
@@ -1202,7 +1203,7 @@ static void LoadPlayerData(game_state* GameState, i32 Handle = -1, glm::vec2 Pos
             }
             else if(StartsWith(&LineBuffer[0], "stamina "))
             {
-                sscanf(LineBuffer, "stamina %hd", &Entity->Player.FullStamina);
+                sscanf(LineBuffer, "stamina %d", &Entity->Player.FullStamina);
                 Entity->Player.Stamina = Entity->Player.FullStamina;
             }
             else if(StartsWith(&LineBuffer[0], "staminagaintimer "))
@@ -1213,19 +1214,19 @@ static void LoadPlayerData(game_state* GameState, i32 Handle = -1, glm::vec2 Pos
             }
             else if(StartsWith(&LineBuffer[0], "hitstaminacost"))
             {
-                sscanf(LineBuffer, "hitstaminacost %hd", &Entity->Player.HitStaminaCost);
+                sscanf(LineBuffer, "hitstaminacost %d", &Entity->Player.HitStaminaCost);
             }
             else if(StartsWith(&LineBuffer[0], "rollstaminacost"))
             {
-                sscanf(LineBuffer, "rollstaminacost %hd", &Entity->Player.RollStaminaCost);
+                sscanf(LineBuffer, "rollstaminacost %d", &Entity->Player.RollStaminaCost);
             }
             else if(StartsWith(&LineBuffer[0], "attackstaminacost"))
             {
-                sscanf(LineBuffer, "attackstaminacost %hd", &Entity->Player.AttackStaminaCost);
+                sscanf(LineBuffer, "attackstaminacost %d", &Entity->Player.AttackStaminaCost);
             }
             else if(StartsWith(&LineBuffer[0], "mindiffstamina"))
             {
-                sscanf(LineBuffer, "mindiffstamina %hd", &Entity->Player.MinDiffStamina);
+                sscanf(LineBuffer, "mindiffstamina %d", &Entity->Player.MinDiffStamina);
             }
             else if(StartsWith(&LineBuffer[0], "staminagaincooldowntimer"))
             {
@@ -1261,9 +1262,9 @@ static void LoadPlayerData(game_state* GameState, i32 Handle = -1, glm::vec2 Pos
     }
 }
 
-static void DecreaseStamina(entity* Entity, game_state* GameState, i16 Cost) 
+static void DecreaseStamina(entity* Entity, game_state* GameState, i32 Cost) 
 {
-    i16 NewStamina = Max(0, Entity->Player.Stamina - Cost);
+    i32 NewStamina = Max(0, Entity->Player.Stamina - Cost);
     Entity->Player.StaminaLost = Entity->Player.Stamina - NewStamina;
     Entity->Player.Stamina = NewStamina;
     StartTimer(GameState, Entity->Player.StaminaDecreaseTimer);
